feature: Completed CRUD cycle with Note Update functionality
Now users can edit the chosen notes with the help of textbox. Edited message can't be empty. User can choose to either save the edit or cancel it. If Save button pressed, POST request is sent to backend and corresponding POST UPDATE handler will update the content of the note, with the timestamp of current edit. Successful note updatewill be shown with the green pop-up message.